add feature migration and related

+++ backend/src/main/java/com/logumrh/service/UserService.java
@@ -31,7 +31,7 @@ public class UserService {
         return userRepository.findByEmail(email);
     }
 
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
         userRepository.deleteById(id);
     }
 }
